grant access to publish event
