+++ cdk/settings/lambda-with-dynamo-db-settings.ts
@@ -1,0 +1,3 @@
+export interface LambdaWithDynamoDbSettings {
+    AWS_NODEJS_CONNECTION_REUSE_ENABLED: string;
+}
