+++ cdk/settings/system-lambda-settings.ts
@@ -1,0 +1,5 @@
+import { LambdaWithDynamoDbSettings } from './lambda-with-dynamo-db-settings';
+
+export interface SystemLambdaSettings extends LambdaWithDynamoDbSettings {
+    SYSTEM_TABLE_NAME: string;
+}
