+++ cdk/settings/use-system-event-bus-lambda-settings.ts
@@ -1,0 +1,3 @@
+export interface UseSystemEventBusLambdaSettings {
+    SYSTEM_EVENT_BUS_NAME: string;
+}
